Fix types in adapter and dashboard
