+++ packages/snap/@types/filecoin-signing-tools/index.d.ts
@@ -19,7 +19,7 @@ declare module "@zondax/filecoin-signing-tools/js" {
         from: string;
         nonce: number;
         value: string;
-        gasprice: string
+        gasprice: string;
         gaslimit: number;
         method: number;
         params?: any;
